+++ ChoGGi's Library/Changelog.md
@@ -1,6 +1,10 @@
 ## Library Changelog
 
 ## v5.4 (Unreleased)
+- Added a bit of colour to the dialogs.
+- I shrunk the Colour Modifier width an update or two ago.
+- Toggling titlebar rollup on a list with hidden ok/cancel buttons stays hidden.
+- Made dialog buttons/checkboxes bold text.
 
 ## v5.3 (08 Feb 2019)
 - List dialog will set width by width of list items.